* 调整updateGroupMyList：数据解析改为在GroupMessageProxy中进行 * 修复部分gid的数据类型，统一成long long * 增加获取过期时间的静态方法：+ (long long)getExpireTime:(NSInteger)minutes;

+++ IMiPhone/DPGroup.h
@@ -12,7 +12,7 @@
 
 @interface DPGroup : NSObject
 #pragma mark - 群基信息
-@property (nonatomic) long  gid;
+@property (nonatomic) long long gid;
 @property (nonatomic, retain) NSString *name;
 @property (nonatomic) NSInteger memberNum;  //成员数
 @property (nonatomic) NSInteger myRelation;
